Address feedback. Fix checkstyle issues.
