+++ gnd/src/main/java/com/google/android/gnd/ui/home/AddFeatureDialogFragment.java
@@ -35,7 +35,7 @@ import com.google.android.gnd.rx.Loadable;
 import com.google.android.gnd.system.auth.AuthenticationManager;
 import com.google.android.gnd.ui.common.AbstractDialogFragment;
 import com.google.android.gnd.ui.home.mapcontainer.MapContainerViewModel;
-import com.google.android.gnd.ui.map.GroundCameraPosition;
+import com.google.android.gnd.ui.map.CameraPosition;
 import com.google.common.collect.ImmutableList;
 import dagger.hilt.android.AndroidEntryPoint;
 import io.reactivex.Maybe;
@@ -81,7 +81,7 @@ public class AddFeatureDialogFragment extends AbstractDialogFragment {
       Project activeProject =
           Loadable.getValue(homeScreenViewModel.getActiveProject())
               .orElseThrow(() -> new NullPointerException("No active project"));
-      GroundCameraPosition cameraPosition =
+      CameraPosition cameraPosition =
           Objects.requireNonNull(
               mapContainerViewModel.getCameraPosition().getValue(), "No camera position");
       return createDialog(activeProject, cameraPosition.getTarget());
