+++ gnd/src/main/java/com/google/android/gnd/ui/map/CameraPosition.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.android.gnd.ui.map;
+
+import com.google.android.gnd.model.feature.Point;
+
+public class CameraPosition {
+
+  private final Point target;
+  private final Float zoomLevel;
+
+  public CameraPosition(Point target, Float zoomLevel) {
+    this.target = target;
+    this.zoomLevel = zoomLevel;
+  }
+
+  public Point getTarget() {
+    return target;
+  }
+
+  public Float getZoomLevel() {
+    return zoomLevel;
+  }
+
+  public String toString() {
+    return "Position: " + target + " Zoom level: " + zoomLevel;
+  }
+}
